+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+-   Adds two new commands, `renpyWarp.lint` and `renpyWarp.rmpersistent`, which
+    allow you to run the Ren'Py linter and delete persistent data respectively
 -   Some older versions of Ren'Py, ~7.4 are now supported
 -   Fix a bug where the extension would be stuck in a 'starting' state if you
     cancel out of the extensions preference prompt